feat: run minion with multiprocessing.Process
* run minion with `multiprocessing.Process` instead of shelling out and
starting the process so we can get logs from it on the console

* break config/swarm bits out into their own files so things are easier
to look at

* remove six and some other python2/3 compatibility poop

BREAKING CHANGE:

Legion is now only compatible with python versions >=3.6
