+++ requirements.txt
@@ -1,0 +1,18 @@
+#
+# This file is autogenerated by pip-compile
+# To update, run:
+#
+#    pip-compile requirements.in
+#
+certifi==2019.11.28       # via requests
+chardet==3.0.4            # via requests
+idna==2.9                 # via requests
+jinja2==2.11.1            # via salt
+markupsafe==1.1.1         # via jinja2, salt
+msgpack==0.6.2            # via salt
+pycrypto==2.6.1           # via salt
+pyyaml==5.3.1             # via salt
+pyzmq==19.0.0             # via salt
+requests==2.23.0          # via salt
+salt==3000                # via -r requirements.in
+urllib3==1.25.8           # via requests
