+++ run.py
@@ -1,0 +1,4 @@
+from legion.legion import main
+
+if __name__ == "__main__":
+    main()
