+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import setup
 setup(
     name="legion",
     version_format="{tag}.dev{commitcount}+{gitsha}",
-    packages=["legion", "legion/modules", "legion/returners",],
+    packages=["legion", "legion/modules", "legion/returners"],
     setup_requires=["setuptools-git-version"],
     entry_points={
         "console_scripts": ["legion=legion.legion:main",],
